Allow using the mouse inside the keyboard
This fires glasses_* events, which are equivalent to mouse-* events but
operating on the glasses' dimensions.

The system is designed to make it trivial to add support for normal
terminals too (firing mouse_* events for that terminal's dimensions).
However, I wasn't entirely happy with the granularity it provided and so
that feature has been left out for now.

I think this closes #108.

+++ src/main/java/org/squiddev/plethora/gameplay/modules/glasses/CanvasHandler.java
@@ -24,8 +24,8 @@ import static net.minecraftforge.common.util.Constants.NBT;
 import static org.squiddev.plethora.gameplay.neural.ItemComputerHandler.MODULE_DATA;
 
 public class CanvasHandler {
-	public static final double WIDTH = 512;
-	public static final double HEIGHT = 512 / 16 * 9;
+	public static final int WIDTH = 512;
+	public static final int HEIGHT = 512 / 16 * 9;
 
 	private static AtomicInteger id = new AtomicInteger(0);
 	private static final HashSet<CanvasServer> server = Sets.newHashSet();
